refactor Posted and rename Assigned to Offered
