--- src/pages/Profile/ProfileContent/Tasks/Assigned/index.js
@@ -1,1 +1,0 @@
-export { default } from './Assigned.jsx';
