+++ src/pages/Profile/ProfileContent/Tasks/Offered/index.js
@@ -1,0 +1,1 @@
+export { default } from './Offered.jsx';
