+++ src/pages/Profile/ProfileContent/Tasks/Posted/TaskModal/index.js
@@ -1,0 +1,1 @@
+export { default } from './TaskModal';
